Add tests & fixes
