+++ device/document/elements.d.ts
@@ -1,0 +1,217 @@
+type ElementState =
+	| 'enabled'
+	| 'disabled'
+	| 'highlighted'
+	| 'unhighlighted'
+	| 'selected'
+	| 'unselected'
+	| 'expanded'
+	| string;
+type ElementSearchMap = {
+	circle: CircleElement;
+	text: TextElement;
+};
+interface ElementSearch {
+	getElementById(id: string): Element;
+	getElementsByClassName(className: string): Element[];
+	getElementsByTagName<TagName extends keyof ElementSearchMap>(
+		tagName: TagName,
+	): Array<ElementSearchMap[TagName]>;
+}
+interface Element extends ElementSearch, GlobalEvents {
+	readonly class: string;
+	readonly firstChild: Element | null;
+	readonly id: string;
+	readonly image: string;
+	layer: number;
+	mode: number | [number, number];
+	readonly nextSibling: Element | null;
+	readonly parent: Element | null;
+	state: ElementState;
+	text: string;
+	readonly type: string;
+	value: number | [number, number];
+	animate(trigger: string): void;
+	sendEvent(event: Event): void;
+	enabled: boolean;
+}
+interface ScrollIndicatorElement extends ContainerElement {
+	length: number;
+	setPosition(position: number, pageSize: number, viewSize: number): void;
+}
+interface ComboButtonEnableDisableOptions {
+	animate?: boolean;
+}
+interface ComboButtonShowHideOptions {
+	animate?: boolean;
+}
+interface ComboButton extends ContainerElement {
+	disable(options?: ComboButtonEnableDisableOptions): void;
+	enable(options?: ComboButtonEnableDisableOptions): void;
+	hide(options?: ComboButtonShowHideOptions): void;
+	show(options?: ComboButtonShowHideOptions): void;
+}
+interface VirtualTileListItemUpdateOptions {
+	redraw?: boolean;
+}
+interface VirtualTileListDelegate {
+	configureTile(
+		tile: VirtualTileListItem,
+		tileInfo: VirtualTileListItemInfo,
+	): void;
+	getTileInfo(position: number): VirtualTileListItemInfo;
+}
+interface VirtualTileListItemInfo {
+	type: string;
+	[extraProperties: string]: any;
+}
+interface VirtualTileList extends ContainerElement {
+	delegate: VirtualTileListDelegate;
+	readonly firstVisibleTile: number;
+	readonly lastVisibleTile: number;
+	length: number;
+	redraw(): void;
+	updateTile(
+		position: number,
+		options?: VirtualTileListItemUpdateOptions,
+	): void;
+}
+type TileListItemAlignment = 'middle' | 'top' | 'bottom';
+interface TileListItemOperationOptions {
+	alignment?: TileListItemAlignment;
+	animate?: boolean;
+	redraw?: boolean;
+}
+interface VirtualTileListItem extends ContainerElement {
+	align(options: TileListItemOperationOptions): void;
+	hide(options?: TileListItemOperationOptions): void;
+	show(options?: TileListItemOperationOptions): void;
+}
+interface TileListItem extends ContainerElement {
+	align(options: TileListItemOperationOptions): void;
+	hide(options?: TileListItemOperationOptions): void;
+	remove(options?: TileListItemOperationOptions): void;
+	show(options?: TileListItemOperationOptions): void;
+}
+interface GraphicsElement extends Element, Styled, Bounded {
+	getBBox(): DOMRect;
+}
+interface TileList extends GraphicsElement {
+	readonly length: number;
+	getTile(position: number): TileListItem;
+	insertTile(
+		item: TileListItem,
+		position: number,
+		options?: TileListItemOperationOptions,
+	): void;
+	obtainTile(type: string): TileListItem | null;
+	redraw(): void;
+}
+interface ImageElement extends GraphicsElement {
+	href: string;
+}
+interface LineElement extends GraphicsElement {
+	x1: number;
+	x2: number;
+	y1: number;
+	y2: number;
+}
+interface GradientArcElement extends ArcElement {
+	readonly gradient: Gradient;
+}
+interface ArcElement extends GraphicsElement {
+	arcWidth: number;
+	startAngle: number;
+	sweepAngle: number;
+}
+interface CircleElement extends GraphicsElement {
+	cx: number;
+	cy: number;
+	r: number;
+}
+interface GradientRectElement extends RectElement {
+	gradient: Gradient;
+}
+interface RectElement extends GraphicsElement, Bounded {}
+interface TextAreaElement extends TextElement {
+	baseline: number;
+	displayAlign: 'top' | 'center' | 'bottom';
+	lineIncrement: number;
+	rows: number;
+}
+interface TextElement extends GraphicsElement {
+	letterSpacing: number;
+	textAnchor: 'start' | 'middle' | 'end';
+	textContent: string;
+	textOverflow: 'clip' | 'ellipsis';
+	readonly textOverflowing: boolean;
+	writingMode: 'lr' | 'rl';
+}
+interface GroupElement extends GraphicsElement {
+	readonly groupTransform: Transform | undefined;
+}
+interface ContainerElement extends GraphicsElement {
+	overflow: 'inherit' | 'visible' | 'hidden';
+}
+interface DOMRect {
+	readonly bottom: number;
+	readonly height: number;
+	readonly left: number;
+	readonly right: number;
+	readonly top: number;
+	readonly width: number;
+	readonly x: number;
+	readonly y: number;
+}
+interface Bounded {
+	height: number;
+	width: number;
+	x: number;
+	y: number;
+}
+interface Styled {
+	readonly style: Style;
+}
+interface Gradient {
+	readonly colors: {
+		c1: string;
+		c2: string;
+		c3: string;
+		c4: string;
+	};
+	opacity: {
+		o1?: number;
+		o2?: number;
+		o3?: number;
+		o4?: number;
+	};
+	type: 'linear' | 'radial' | 'bilinear';
+	x1: number;
+	y1: number;
+	x2: number;
+	y2: number;
+}
+interface Transform {
+	readonly rotate: {
+		angle: number;
+	};
+	readonly scale: {
+		x: number;
+		y: number;
+	};
+	readonly translate: {
+		x: number;
+		y: number;
+	};
+}
+interface Style {
+	display: 'none' | 'inline' | 'inherit';
+	fill: string;
+	fillOpacity: number | undefined;
+	fontFamily: string;
+	fontSize: number | undefined;
+	fontWeight: 'inherit' | 'light' | 'regular' | 'bold';
+	opacity: number | undefined;
+	textDecoration: 'none' | 'inherit' | 'underline';
+	visibility: 'inherit' | 'visible' | 'hidden';
+}
