+++ settings/shared.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="react" />
 /// <reference path="../shared/globals.d.ts" />
 /// <reference path="../shared/console.d.ts" />
 /// <reference path="../shared/event.d.ts" />
