+++ test/device/user-interface/activate-event.ts
@@ -1,0 +1,7 @@
+import document from 'document';
+
+let mybutton = document.getElementById('mybutton');
+
+mybutton.onactivate = function(e) {
+	console.log('click');
+};
