+++ test/device/user-interface/enable-disable.ts
@@ -1,0 +1,5 @@
+import document from 'document';
+
+let mybutton = document.getElementById('mybutton');
+mybutton.enabled = false; // disabled
+mybutton.enabled = true; // enabled
