+++ test/device/user-interface/events.ts
@@ -1,0 +1,7 @@
+import document from 'document';
+
+let myRect = document.getElementById('myRect');
+
+myRect.onclick = function(e) {
+	console.log('click');
+};
