+++ test/device/user-interface/image.ts
@@ -1,0 +1,4 @@
+import document from 'document';
+
+let img = document.getElementById('image') as ImageElement;
+img.href = 'another-image.jpg';
