+++ test/device/user-interface/layer-stacking.ts
@@ -1,0 +1,4 @@
+import document from 'document';
+
+let demo = document.getElementById('demo') as GraphicsElement;
+demo.layer = 2;
