+++ test/device/user-interface/marquee-text.ts
@@ -1,0 +1,7 @@
+import document from 'document';
+
+let marquee = document.getElementById('marquee');
+
+setTimeout(function() {
+	marquee.state = 'enabled';
+}, 2000);
