+++ test/device/user-interface/move-an-element.ts
@@ -1,0 +1,5 @@
+import document from 'document';
+
+let demo = document.getElementById('demo') as GraphicsElement;
+demo.x = 42;
+demo.y = 90;
