+++ test/device/user-interface/physical-buttons.ts
@@ -1,0 +1,5 @@
+import document from 'document';
+
+document.onkeypress = function(e) {
+	console.log('Key pressed: ' + e.key);
+};
