+++ test/device/user-interface/resize-an-element.ts
@@ -1,0 +1,5 @@
+import document from 'document';
+
+let demo = document.getElementById('demo') as GraphicsElement;
+demo.width = 348;
+demo.height = 250;
