+++ test/device/user-interface/spinner.ts
@@ -1,0 +1,11 @@
+import document from 'document';
+
+let spinner = document.getElementById('spinner');
+
+// Start the spinner
+spinner.state = 'enabled';
+
+setTimeout(function() {
+	// Stop the spinner
+	spinner.state = 'disabled';
+}, 4000);
