+++ test/device/user-interface/sprite-image.ts
@@ -1,0 +1,6 @@
+import document from 'document';
+
+let myanimation = document.getElementById('myanimation');
+
+myanimation.animate('enable');
+myanimation.animate('disable');
