+++ test/device/user-interface/styling.ts
@@ -1,0 +1,6 @@
+import document from 'document';
+
+let demo = document.getElementById('demo') as GraphicsElement;
+demo.style.fill = 'red';
+demo.style.fontSize = 20;
+demo.style.fontFamily = 'System-Regular';
