+++ test/tsconfig.json
@@ -3,6 +3,7 @@
 		"target": "es5",
 		"module": "commonjs",
 		"noEmit": true,
-		"lib": ["es2015"]
+		"lib": ["es2015"],
+		"typeRoots": []
 	}
 }
